+++ week4/src/types/form.ts
@@ -1,0 +1,7 @@
+import { SignupData } from '@type/user';
+import { ChangeEvent } from 'react';
+
+export interface SingupFormData {
+  formData: SignupData;
+  onInputChange: (e: ChangeEvent<HTMLInputElement>) => void;
+}
